requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.3.6
redis==5.0.1
django-celery-Beat==2.5.0